DEV: Replace `params` by the contract attributes in services
This patch allows accessing a sanitized version of parameters provided
to a service through `params` once the service contract has run.

The whole contract is still available through `contract`, but now its
attributes automatically erase those in `params`.

That way, it allows better consistency when accessing input params. For
example, if you have a service without a contract, to access a
parameter, you need to use `params[:my_parameter]`. But with a contract,
you do this through `contract.my_parameter`. Now with this patch, you’ll
be able to do `params[:my_parameter]` to access a sanitized value
without having to give it too much thinking.